[#115362749] int->float in conversion parameters builder

+++ ambassadorsdk/src/main/java/com/ambassador/ambassadorsdk/ConversionParameters.java
@@ -154,7 +154,7 @@ public final class ConversionParameters {
         private String mbsy_custom2;
         private String mbsy_custom3;
         private int mbsy_auto_create;
-        private int mbsy_revenue = -1;
+        private float mbsy_revenue = -1;
         private int mbsy_deactivate_new_ambassador;
         private String mbsy_transaction_uid;
         private String mbsy_add_to_group_id;
@@ -232,7 +232,7 @@ public final class ConversionParameters {
         }
 
         @NonNull
-        public Builder setRevenue(int mbsy_revenue) {
+        public Builder setRevenue(float mbsy_revenue) {
             this.mbsy_revenue = mbsy_revenue;
             this.revenueCheck = true;
             return this;
